api/pyproject.toml
[project]
name = "vj-api"
version = "1.0.6"
description = "Provides a random YouTube video ID according to an optional given theme, and store cached YouTube IDs in a DB for later usage without depleting the YouTube API quota."
authors = [{ name = "Adrien Carpentier", email = "me@adriencarpentier.com" }]
dependencies = [
    "colorlog>=6.7.0",
    "django>=5.1",
    "django-ninja>=0.22.2",
    "django-cors-headers>=4.2.0",
    "gunicorn>=21.2.0",
    "langdetect>=1.0.9",
    "nltk>=3.8",
    "orjson>=3.9.5",
    "psycopg-binary>=3.1.10",
    "psycopg>=3.1.10",
    "requests>=2.31.0",
    "sentry-sdk>=1.29.2",
    "uvicorn>=0.23.2",
]
requires-python = ">=3.11,<3.14"
dynamic = ["classifiers", "classifiers"]
license = { text = "MIT" }

[project.urls]
homepage = "https://vj.adriencarpentier.com/api"

[tool.ruff]
line-length = 100
lint = { extend-select = ["I"] }

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.7",
]
